refactor(feedback): move Aperture logic out of Core into the handler
Feedback is an outbound product-API call, not work in the user's AWS account, so
it no longer belongs on the AWS-focused Core surface (review: AlexanderRichey).
Moves the submit logic to src/handlers/feedback/submit.ts, injected with a fetch
threaded through RootHandlerConfig (defaults to global fetch; tests inject their
own), and drops Core.feedback / CoreClient.feedback / TestFeedbackClient and
src/core/feedback.tsx.

Folds in the review fixes to the moved code:
- Derive+validate the S3 object key BEFORE the PUT, so an invalid presign body
  raises ApertureError instead of a bare TypeError inside the upload (aidandaly24).
- Validate the form response shape (string id/timestamp/reference) instead of an
  unchecked cast; a partial/non-JSON 2xx now raises ApertureError (harness bot).
- Inline the over-defensive readBody into `response.text().catch(() => "")`
  (AlexanderRichey).
- Feedback data shapes (ScreenshotInput/SubmitFeedbackInput/FeedbackSubmissionResult
  and internal records) are `type` aliases; `interface` is reserved for contracts
  (aidandaly24).
