+++ src/handlers/feedback/types.tsx
@@ -1,26 +1,16 @@
-import type { CoreOptions } from "../../core/types";
+// Concrete request/result data shapes for the feedback command. These are data,
+// not behavioral contracts, so they are type aliases (interfaces are reserved for
+// the Core*Client contracts elsewhere in handlers/).
 
-export interface ScreenshotInput {
-  path: string;
-}
+export type ScreenshotInput = { path: string };
 
-export interface SubmitFeedbackInput {
+export type SubmitFeedbackInput = {
   message: string;
   screenshot?: ScreenshotInput;
-}
+};
 
-export interface FeedbackSubmissionResult {
+export type FeedbackSubmissionResult = {
   id: string;
   timestamp: string;
   reference: string;
-}
-
-// Consumer-defined interface (dependency inversion): the handler depends on this,
-// src/core/feedback.tsx implements it. Message/screenshot validation happens inside
-// submitFeedback so the one code path guards every caller.
-export interface CoreFeedbackClient {
-  submitFeedback(
-    input: SubmitFeedbackInput,
-    options: CoreOptions,
-  ): Promise<FeedbackSubmissionResult>;
-}
+};
